fix conflicts after #546
