+++ frontend/app/utils/utilts.store.ts
@@ -1,6 +1,9 @@
 import { deepEqual } from 'fast-equals'
 import { snapshot, subscribe } from 'valtio'
-import { createDefaultBannerProfile } from '@shared/default-data'
+import {
+  createDefaultBannerProfile,
+  createDefaultWidgetProfile,
+} from '@shared/default-data'
 import {
   type ProfileId,
   type Tool,
@@ -9,10 +12,12 @@ import {
   DEFAULT_PROFILE_NAMES,
   PROFILE_IDS,
   TOOL_BANNER,
+  TOOL_WIDGET,
 } from '@shared/types'
 import type { BannerStore } from '~/stores/banner-store'
+import type { WidgetStore } from '~/stores/widget-store'
 
-type Store = BannerStore
+type Store = BannerStore | WidgetStore
 const STORAGE_PREFIX = 'wmt'
 
 export function getStorageKeys(tool: Tool) {
@@ -27,6 +32,8 @@ function getCreateDefaultProfile<T extends Tool>(tool: T) {
   switch (tool) {
     case TOOL_BANNER:
       return createDefaultBannerProfile
+    case TOOL_WIDGET:
+      return createDefaultWidgetProfile
     default:
       throw new Error(`Unknown tool: ${tool}`)
   }
